// src/app/shares/models/applicant.mode.ts
export interface applicant {
    id: Number;
    name: String;
    ustEmail: String;
    year: String;
    section: String;
    studentNo: String;
    sex: String;
    birthDate: Date;
    age: number;
    membershipStatus: String;
    contact: number;
    address: String;
    facebookLink: String;
    lookingForward: String;
    typeOfEvents: String;
    topicsIntrested: String;
}